pyproject.toml
[tool.black]
# line-length = 119
target-version = ['py39']

[tool.mypy]
python_version = "3.9"

[tool.ruff]
# Never enforce `E501` (line length violations).
ignore = ["C901", "E501", "E741", "F405", "F403", "W605"]
# select = ["C", "E", "F", "I", "W"]

# # Ignore import violations in all `__init__.py` files.
# [tool.ruff.per-file-ignores]
# "__init__.py" = ["E402", "F401", "F403", "F811"]

# [tool.ruff.isort]
# lines-after-imports = 2

[tool.poetry]
name = "lrp-for-resnet"
version = "0.1.0"
description = ""
authors = ["Foo <foo@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
torch = "2.0.0"
torchvision = "0.15.1"
torchaudio = "2.0.1"
tqdm = "^4.65.0"
matplotlib = "^3.7.1"
numpy = "^1.24.3"
opencv-python = "^4.7.0.72"
pillow = "^9.5.0"
scikit-image = "^0.20.0"
scikit-learn = "^1.2.2"
torchinfo = "^1.8.0"
wandb = "^0.15.3"
timm = "^0.9.2"
grad-cam = "^1.4.6"
captum = "^0.6.0"
plotly = "^5.19.0"
kaleido = "0.2.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.269"
black = "^23.3.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
